2025/7_speech_to_code/requirements.txt
assemblyai>=0.10.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
rich>=13.0.0